rename mixin folder to mixins
