+++ frontend/src/mixins/getCommunityInfo.js
